ISPN-13648 Isolate legacy annotations access
