+++ remote-query/remote-query-server/src/main/java/org/infinispan/query/remote/impl/ProtobufMetadataManagerImpl.java
@@ -46,7 +46,7 @@ import org.infinispan.protostream.types.java.CommonTypesSchema;
 import org.infinispan.query.remote.ProtobufMetadataManager;
 import org.infinispan.query.remote.client.ProtobufMetadataManagerConstants;
 import org.infinispan.query.remote.client.impl.MarshallerRegistration;
-import org.infinispan.query.remote.impl.indexing.IndexingMetadata;
+import org.infinispan.query.remote.impl.indexing.search5.Search5Annotations;
 import org.infinispan.query.remote.impl.logging.Log;
 import org.infinispan.registry.InternalCacheRegistry;
 import org.infinispan.security.impl.CreatePermissionConfigurationBuilder;
@@ -80,7 +80,7 @@ public final class ProtobufMetadataManagerImpl implements ProtobufMetadataManage
 
    public ProtobufMetadataManagerImpl() {
       Configuration.Builder protostreamCfgBuilder = Configuration.builder();
-      IndexingMetadata.configure(protostreamCfgBuilder);
+      Search5Annotations.configure(protostreamCfgBuilder);
       serCtx = ProtobufUtil.newSerializationContext(protostreamCfgBuilder.build());
       try {
          MarshallerRegistration.init(serCtx);
