+++ remote-query/remote-query-server/src/main/java/org/infinispan/query/remote/impl/mapping/reference/MessageReferenceProvider.java
@@ -1,8 +1,5 @@
 package org.infinispan.query.remote.impl.mapping.reference;
 
-import static org.infinispan.query.remote.impl.indexing.IndexingMetadata.FIELD_ANNOTATION;
-import static org.infinispan.query.remote.impl.indexing.IndexingMetadata.FIELD_INDEX_ATTRIBUTE;
-import static org.infinispan.query.remote.impl.indexing.IndexingMetadata.findAnnotation;
 import static org.infinispan.query.remote.impl.indexing.IndexingMetadata.findProcessedAnnotation;
 
 import java.util.ArrayList;
@@ -11,7 +8,6 @@ import java.util.List;
 
 import org.hibernate.search.engine.backend.document.IndexFieldReference;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement;
-import org.infinispan.protostream.descriptors.AnnotationElement;
 import org.infinispan.protostream.descriptors.Descriptor;
 import org.infinispan.protostream.descriptors.FieldDescriptor;
 import org.infinispan.protostream.descriptors.Type;
@@ -32,7 +28,7 @@ public class MessageReferenceProvider {
       this.fields = new ArrayList<>(descriptor.getFields().size());
       this.embedded = new ArrayList<>();
 
-      IndexingMetadata indexingMetadata =  findProcessedAnnotation(descriptor, IndexingMetadata.INDEXED_ANNOTATION);
+      IndexingMetadata indexingMetadata = findProcessedAnnotation(descriptor, IndexingMetadata.INDEXED_ANNOTATION);
       // Skip if not annotated with @Indexed
       if (indexingMetadata == null) {
          return;
@@ -40,19 +36,19 @@ public class MessageReferenceProvider {
 
       for (FieldDescriptor fieldDescriptor : descriptor.getFields()) {
          String fieldName = fieldDescriptor.getName();
-         if (Type.MESSAGE.equals(fieldDescriptor.getType())) {
-            // If a protobuf field is a Message reference, only take it into account
-            // if the Message is @Indexed and has at least one @Field annotation
-            if (indexEmbedded(fieldDescriptor)) {
-               // Hibernate Search can handle the @Field regardless of its attributes
-               embedded.add(new Embedded(fieldName, fieldDescriptor.getMessageType().getFullName(), fieldDescriptor.isRepeated()));
-            }
-            continue;
-         }
-
          FieldMapping fieldMapping = indexingMetadata.getFieldMapping(fieldName);
          if (fieldMapping == null) {
             // field model is not mapped
+            continue;
+         }
+
+         if (Type.MESSAGE.equals(fieldDescriptor.getType())) {
+            // If a protobuf field is a Message reference, only take it into account
+            // if the Message is @Indexed and has at least one @Field annotation
+            if (fieldMapping.index() && isIndexable(fieldDescriptor.getMessageType())) {
+               // Hibernate Search can handle the @Field regardless of its attributes
+               embedded.add(new Embedded(fieldName, fieldDescriptor.getMessageType().getFullName(), fieldDescriptor.isRepeated()));
+            }
             continue;
          }
 
@@ -63,27 +59,22 @@ public class MessageReferenceProvider {
       }
    }
 
-   private boolean indexEmbedded(FieldDescriptor fieldDescriptor) {
-      AnnotationElement.Annotation fieldAnnotation = findAnnotation(fieldDescriptor.getAnnotations(), FIELD_ANNOTATION);
-      if (fieldAnnotation == null) {
-         return false;
-      }
-
-      AnnotationElement.Value indexAttribute = fieldAnnotation.getAttributeValue(FIELD_INDEX_ATTRIBUTE);
-      boolean isIndexed = IndexingMetadata.INDEX_YES.equals(indexAttribute.getValue()) || IndexingMetadata.YES.equals(indexAttribute.getValue());
-      if (!isIndexed) {
-         return false;
-      }
-
-      return isIndexable(fieldDescriptor.getMessageType());
-   }
-
    /**
     * Checks if a Descriptors has the @Indexed annotation and at least one @Field
     */
    private boolean isIndexable(Descriptor descriptor) {
-      IndexingMetadata indexingMetadata =  findProcessedAnnotation(descriptor, IndexingMetadata.INDEXED_ANNOTATION);
-      return indexingMetadata != null && descriptor.getFields().stream().anyMatch(f -> findAnnotation(f.getAnnotations(), FIELD_ANNOTATION) != null);
+      IndexingMetadata indexingMetadata = findProcessedAnnotation(descriptor, IndexingMetadata.INDEXED_ANNOTATION);
+      if ( indexingMetadata == null ) {
+         return false;
+      }
+
+      for (FieldDescriptor fieldDescriptor : descriptor.getFields()) {
+         FieldMapping fieldMapping = indexingMetadata.getFieldMapping(fieldDescriptor.getName());
+         if (fieldMapping != null) {
+            return true;
+         }
+      }
+      return false;
    }
 
    public boolean isEmpty() {
